requirements.txt
flask
gunicorn
requests
pdfplumber
python-docx
pydantic
python-multipart
reportlab
PyPDF2
openai
pdfkit
jinja2
python-dotenv
fpdf
PyMuPDF
celery
redis

